:recycle: (ControlComponent) Move in ticket feature where it is used
